+++ ItemQualities/Assets/ItemQualities/CHANGELOG.md
@@ -6,6 +6,10 @@
     * <details><summary>Safer Spaces</summary> On block, apply a fog to nearby enemies that increases damage from all void-based attacks.</details>
     * <details><summary>The Crowdfunder</summary> Bullets can ricochet to nearby enemies. Increased max fire rate (same stats as previous version)</details>
 
+* Deus Ex Machina:
+    * The stored projectile is now fired by using your primary skill, rather than next equipment activation.
+    * Stored projectile now persists between stages.
+
 * Preon Accumulator:
     * Explosion/zap radius:
         * Uncommon: +5m -> +10m
@@ -38,6 +42,13 @@
         * Legendary: 15s -> 20s
     * Base cooldown (for all quality variants): 15s -> 30s
         * This was done to allow the sprout to remain for longer while still having a cooldown after the sprout times out.
+
+* Ceremonial Dagger:
+    * Additional dagger amount:
+        * Uncommon: 1 (+1 per stack) -> 2 (+2 per stack)
+        * Rare: 2 (+2 per stack) -> 5 (+5 per stack)
+        * Epic: 5 (+5 per stack) -> 8 (+8 per stack)
+        * Legendary: 7 (+7 per stack) -> 10 (+10 per stack)
 
 * Gnarled Woodsprite:
     * Cloned Junk Drone temp item duration: 100% -> 25%